+++ libearth/subscribe.py
@@ -284,11 +284,10 @@ class SubscriptionSet(collections.MutableSet):
                     elif outline.deleted_at:
                         child.deleted_at = max(child.deleted_at,
                                                outline.deleted_at)
-                    if child.children:
-                        if child.deleted:
-                            del child.children[:]
-                        else:
-                            child.__merge_entities__(outline)
+                    if child.children and child.deleted:
+                        del child.children[:]
+                    if child.children or outline.children:
+                        SubscriptionSet.__merge_entities__(child, outline)
                     break
             else:
                 self.add(outline)